training: RAM-bounded shard loader + resume support
Rewrites TrainingDataset to lazy-load shards through a 4-shard LRU instead
of holding all 569 shard payloads in memory. Adds ShardChunkSampler so the
cache hits, split_by_shard for shard-granular train/val, and --resume to
recover from container kills via per-epoch last.pt. Shrinks the volatile
bitmask from float32 to bool (promoted at projection time) to keep shards
small as the corpus grows.

Unblocks v8 corpus refresh and any future expansion past the 1200+ rating
filter without OOMing the host.

+++ src/vgc_model/model/action_model.py
@@ -256,7 +256,8 @@ class ActionModel(nn.Module):
         if self.use_boosts:
             slots = slots + self.boost_proj(batch["stat_boosts"])
         if self.use_volatile:
-            slots = slots + self.volatile_proj(batch["volatiles"])
+            # Volatile bitmask is stored bool to save shard RAM; promote to float.
+            slots = slots + self.volatile_proj(batch["volatiles"].float())
         if self.use_sub_hp:
             slots = slots + self.sub_hp_proj(batch["sub_hps"].unsqueeze(-1))
